+++ src/Console/Commands/HealthPanelCommand.php
@@ -22,10 +22,12 @@ class HealthPanelCommand extends Command
     protected $description = 'Show all resources and their current health states.';
 
     /**
-     * @return void
+     * @return int Exit code: 0 = success; 2 = failed
      */
-    public function handle(Commands $commands): void
+    public function handle(Commands $commands): int
     {
-        $commands->panel($this);
+        $this->info('Checking resources and gathering information to build the panel...');
+
+        return $commands->panel($this);
     }
 }