+++ src/Http/Controllers/Health.php
@@ -2,10 +2,12 @@
 
 namespace PragmaRX\Health\Http\Controllers;
 
+use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\File;
 use PragmaRX\Health\Service;
+use PragmaRX\Health\Support\Result;
 
 class Health extends Controller
 {
@@ -17,7 +19,7 @@ class Health extends Controller
     /**
      * Health constructor.
      *
-     * @param Service $healthService
+     * @param  Service  $healthService
      */
     public function __construct(Service $healthService)
     {
@@ -28,6 +30,7 @@ class Health extends Controller
      * Check all resources.
      *
      * @return array
+     *
      * @throws \Exception
      */
     public function check()
@@ -42,19 +45,61 @@ class Health extends Controller
      *
      * @param $slug
      * @return mixed
+     *
      * @throws \Exception
      */
-    public function getResource($slug)
+    public function getResource($slug, Request $request)
     {
         $this->healthService->setAction('resource');
 
-        return $this->healthService->resource($slug);
+        $resource = $this->healthService->resource($slug);
+
+        // Get any expected response format (fallsback to JSON)
+        $format = $request->get('format');
+
+        // Summary format - Prints status: resource, and any extra piece of information we have.
+        if ($format === 'summary') {
+            $format = 'M j H:m:s';
+            $now = date($format, time());
+
+            // Additional details to include in the response, such as errors
+            $more = '';
+
+            switch ($resource->getStatus()) {
+                case Result::OK:
+                    $msg = "{$resource->name} is running as expected";
+                    break;
+                case Result::WARNING:
+                    $msg = "{$resource->name} is running above the warning threshold";
+                    break;
+                case Result::CRITICAL:
+                    if (! empty($resource->errorMessage)) {
+                        $msg = "{$resource->errorMessage}";
+                    } else {
+                        $msg = "{$resource->name} service is failing or has reached the critical threshold";
+                    }
+                    $checkerClassName = (new \ReflectionClass(get_class($resource->checker)))->getName();
+                    $more = "{$checkerClassName} was used to determine the health";
+                    break;
+                case Result::UNKNOWN:
+                default:
+                    $msg = "{$resource->name} service health is unknown";
+                    break;
+            }
+            $response = \strtoupper($resource->getStatus()).": {$msg} (Checked {$now})\n{$more}";
+
+            return response($response)
+                ->header('Content-Type', 'text/plain');
+        } else {
+            return $resource;
+        }
     }
 
     /**
      * Get all resources.
      *
      * @return mixed
+     *
      * @throws \Exception
      */
     public function allResources()
@@ -64,14 +109,17 @@ class Health extends Controller
 
     /**
      * @return mixed
+     *
      * @throws \Exception
      */
-    public function string()
+    public function string(Request $request)
     {
+        $filters = $request->get('filters');
+
         $this->healthService->setAction('string');
 
         return response(
-            $this->healthService->string()
+            $this->healthService->string($filters)
         );
     }
 
@@ -91,6 +139,7 @@ class Health extends Controller
 
     /**
      * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     *
      * @throws \Exception
      */
     public function panel()
@@ -125,5 +174,10 @@ class Health extends Controller
     public function config()
     {
         return config('health');
+    }
+
+    public function serverVars()
+    {
+        return $_SERVER;
     }
 }