+++ src/Support/Constants.php
@@ -5,4 +5,6 @@ namespace PragmaRX\Health\Support;
 class Constants
 {
     const RESOURCES_ENABLED_ALL = ['.*']; // regex
+
+    const SERVER_VARS_CACHE_KEY_PREFIX = 'pragmarx-health-locally-protected';
 }