+++ src/Support/Result.php
@@ -5,6 +5,20 @@ namespace PragmaRX\Health\Support;
 class Result
 {
     /**
+     * States the result of the check could be in
+     * Further reading: https://nagios-plugins.org/doc/guidelines.html#AEN78.
+     */
+    const OK = 'OK';
+    const WARNING = 'Warning';
+    const CRITICAL = 'Critical';
+    const UNKNOWN = 'Unknown';
+
+    /**
+     * Defaults the state of the result to unknown.
+     */
+    protected $state = self::UNKNOWN;
+
+    /**
      * @var bool
      */
     public $healthy;
@@ -34,12 +48,15 @@ class Result
         $this->healthy = $healthy;
 
         $this->errorMessage = $errorMessage;
+
+        // Currently the status is inferred from the $healthy flag until full support is added.
+        $this->status = $healthy ? self::OK : self::CRITICAL;
     }
 
     /**
      * Value setter.
      *
-     * @param mixed $value
+     * @param  mixed  $value
      * @return Result
      */
     public function setValue($value)
@@ -52,7 +69,7 @@ class Result
     /**
      * Value for humans setter.
      *
-     * @param string $valueHuman
+     * @param  string  $valueHuman
      * @return Result
      */
     public function setValueHuman(string $valueHuman)
@@ -60,5 +77,15 @@ class Result
         $this->valueHuman = $valueHuman;
 
         return $this;
+    }
+
+    /**
+     * Get the result's status of the check.
+     *
+     * @return string one of the consts e.g. result::OK
+     */
+    public function getStatus(): string
+    {
+        return $this->status;
     }
 }